move conv1d to adapter; move qwen_utils to examples
