+++ examples/qwen2_5_omni/qwen_omni_utils/vision_process.py
